reverse the order of attempted matches in a qualification
it was trying to match static values first, then the "from … where" form
except that I had told it that the static form couldn't be followed by a
paren; all well and good. except that I didn't tell it that it couldn't
be followed by a space-then-paren.

so i just switch the attempted matching order.